+++ vue-task-tracker-2021/src/App.vue
@@ -5,10 +5,8 @@
       title="Task Tracker"
       :showAddTaskBtn="showAddTask"
     />
-    <div v-if="showAddTask">
-      <AddTask @add-task="addTask" />
-    </div>
-    <Tasks @delete-task="deleteTask" v-bind:tasks="tasks" />
+
+    <router-view :showAddTask="showAddTask"></router-view>
     <Footer />
   </div>
 </template>
@@ -16,66 +14,21 @@
 <script>
 import Header from "./components/Header";
 import Footer from "./components/Footer";
-import Tasks from "./components/Tasks";
-import AddTask from "./components/AddTask";
 export default {
   name: "App", //register components here
   components: {
     Header,
     Footer,
-    Tasks,
-    AddTask,
   },
   data() {
     return {
-      tasks: [],
       showAddTask: false,
     };
   },
   methods: {
-    async addTask(task) {
-      const resq = await fetch("api/tasks", {
-        method: "POST",
-        headers: { "Content-type": "application/json" },
-        body: JSON.stringify(task),
-      });
-
-      var data = await resq.json();
-      this.tasks = [...this.tasks, data];
-    },
-    async deleteTask(task) {
-      if (confirm(`Delete Task: ${task.text}?`)) {
-        const resq = await fetch(`api/tasks/${task.id}`, {
-          method: "DELETE",
-        });
-
-        resq.status === 200
-          ? (this.tasks = this.tasks.filter((x) => x.id !== task.id))
-          : alert(`Error deleting task ${task.text}`);
-      }
-    },
     toggleAddBtn() {
       this.showAddTask = !this.showAddTask;
     },
-    async getTasks() {
-      //populate list on load
-      const response = await fetch("api/tasks");
-
-      const data = await response.json();
-
-      return data;
-    },
-    async getTask(id) {
-      const response = await fetch(`api/tasks/${id}`);
-
-      const data = await response.json();
-
-      return data;
-    },
-  },
-  async created() {
-    //created is a lifecycle method; similar to ngOnInit
-    this.tasks = await this.getTasks();
   },
 };
 </script>
