+++ vue-task-tracker-2021/src/router/index.js
@@ -1,12 +1,19 @@
 import {createRouter, createWebHistory} from 'vue-router'
-import Info from '../views/Info.vue'
+import Info from '../views/Info'
+import Home from '../views/Home'
 
 const routes = [
+    {
+        path: '/',
+        name: 'Home',
+        component: Home
+    },
     {
         path: '/info',
         name: 'Info',
         component: Info
     }
+
 ]
