+++ vue-task-tracker-2021/src/views/Info.vue
@@ -1,3 +1,14 @@
 <template>
-  <h1>Info</h1>
+  <div class="center">
+    <h4>Version 1.0.0</h4>
+    <router-link to="/" class="link">Home</router-link>
+  </div>
 </template>
+
+
+<style scoped>
+.center {
+  text-align: center;
+  padding-top: 200px;
+}
+</style>
